tetris: Use uint8_t instead of int for pieces to save memory

+++ tetris/include/tetris.h
@@ -2,14 +2,15 @@
 #define TETRIS_H
 
 #include <stdbool.h>
+#include <stdint.h>
 #include "settings.h"
 
 enum shape{SquareShape, TShape, ZShape, SShape, JShape, LShape, IShape};
 
 struct pieceInfo{
-    int piece[MAX_PIECE_LEN];
-    int pieceX;
-    int pieceY;
+    uint8_t piece[MAX_PIECE_LEN];
+    uint8_t pieceX;
+    uint8_t pieceY;
     enum shape pieceShape;
 };
 
@@ -17,8 +18,8 @@ enum userRequest{requestLeft, requestRight, requestDrop, requestRotateRight, req
 enum boardAction{none, redraw, moveRight, moveLeft, dropOne, instantDrop};
 
 
-void rotateRight(int *dest, enum shape curShape);
-void rotateLeft(int *dest,  enum shape curShape);
+void rotateRight(uint8_t *dest, enum shape curShape);
+void rotateLeft(uint8_t *dest,  enum shape curShape);
 
 
 void initBoard(int *boardToInit);
@@ -29,9 +30,9 @@ void clearRows(int *boardToClear, int pieceY);
 
 enum boardAction handleUserInput(enum userRequest r, int *board, struct pieceInfo *playerPiece);
 
-bool checkSpawnPiece(int *pieceToSpawn, int *boardToSpawnIn, enum shape pieceShape);
+bool checkSpawnPiece(uint8_t *pieceToSpawn, int *boardToSpawnIn, enum shape pieceShape);
 void renderBoard(int *renderedBoard, int *boardToRender, struct pieceInfo *pieceToRender, bool substract);
-bool checkMove(int *piece, int piecePosX, int piecePosY, enum shape pieceShape, int *boardToCheck);
+bool checkMove(uint8_t *piece, uint8_t piecePosX, uint8_t piecePosY, enum shape pieceShape, int *boardToCheck);
 
 
 #endif
