+++ tetris/src/tetris.c
@@ -5,15 +5,15 @@
 
 #define NUM_SHAPES 7
 
-const int square[]= {0,1,1,0,1,1,0,0,0};
-const int t[] = {1,1,1,0,1,0,0,0,0};
-const int z[] = {1,1,0,0,1,1,0,0,0};
-const int s[] = {0,1,1,1,1,0,0,0,0};
-const int j[] = {1,1,1,0,0,1,0,0,0};
-const int l[] = {1,1,1,1,0,0,0,0,0};
-const int i[] = {0,0,0,0,1,1,1,1,0,0,0,0,0,0,0,0};
+const uint8_t square[]= {0,1,1,0,1,1,0,0,0};
+const uint8_t t[] = {1,1,1,0,1,0,0,0,0};
+const uint8_t z[] = {1,1,0,0,1,1,0,0,0};
+const uint8_t s[] = {0,1,1,1,1,0,0,0,0};
+const uint8_t j[] = {1,1,1,0,0,1,0,0,0};
+const uint8_t l[] = {1,1,1,1,0,0,0,0,0};
+const uint8_t i[] = {0,0,0,0,1,1,1,1,0,0,0,0,0,0,0,0};
 
-const int *figs[NUM_SHAPES] = {square, t, z, s, j, l, i};
+const uint8_t *figs[NUM_SHAPES] = {square, t, z, s, j, l, i};
 
 int pieceLenByShape(enum shape pieceShape){
     if(pieceShape == IShape){
@@ -38,7 +38,7 @@ void initRandomPiece(struct pieceInfo *pieceToInit){
     int shapeNum = rand() % NUM_SHAPES;
     enum shape shapeType = shapeNum;
 
-    memcpy(pieceToInit->piece, figs[shapeNum], pieceLenByShape(shapeType) * sizeof(int));
+    memcpy(pieceToInit->piece, figs[shapeNum], pieceLenByShape(shapeType) * sizeof(uint8_t));
     pieceToInit->pieceShape = shapeType;
 
     pieceToInit->pieceX = SPAWN_X;
@@ -51,7 +51,7 @@ void updateBoard(int *board, struct pieceInfo *playerPiece){
     memcpy(board, renderedBoard, sizeof(int) * BOARDSIZE);
 }
 
-bool checkSpawnPiece(int *pieceToSpawn, int *boardToSpawnIn, enum shape pieceShape){
+bool checkSpawnPiece(uint8_t *pieceToSpawn, int *boardToSpawnIn, enum shape pieceShape){
     return checkMove(pieceToSpawn, SPAWN_X, SPAWN_Y, pieceShape, boardToSpawnIn);
 }
 
@@ -82,7 +82,7 @@ void renderBoard(int *renderedBoard, int *boardToRender, struct pieceInfo *piece
 enum boardAction handleUserInput(enum userRequest r, int *board, struct pieceInfo *playerPiece){
     
     int pieceLen = pieceLenByShape(playerPiece->pieceShape);
-    int modifiedPiece[MAX_PIECE_LEN];
+    uint8_t modifiedPiece[MAX_PIECE_LEN];
 
     int pieceX = playerPiece->pieceX;
     int pieceY = playerPiece->pieceY;
@@ -90,21 +90,21 @@ enum boardAction handleUserInput(enum userRequest r, int *board, struct pieceInf
 
     int yOffset = 0;
 
-    memcpy(modifiedPiece, playerPiece->piece, sizeof(int) * pieceLen);
+    memcpy(modifiedPiece, playerPiece->piece, sizeof(uint8_t) * pieceLen);
 
     switch (r)
     {
     case requestRotateRight:
         rotateRight(modifiedPiece, pieceShape);
         if(checkMove(modifiedPiece, pieceX, pieceY, pieceShape, board)){
-            memcpy(playerPiece->piece, modifiedPiece, sizeof(int) * pieceLen);
+            memcpy(playerPiece->piece, modifiedPiece, sizeof(uint8_t) * pieceLen);
         }
         return redraw;
 
     case requestRotateLeft:
         rotateLeft(modifiedPiece, pieceShape);
         if(checkMove(modifiedPiece, pieceX, pieceY, pieceShape, board)){
-            memcpy(playerPiece->piece, modifiedPiece, sizeof(int) * pieceLen);
+            memcpy(playerPiece->piece, modifiedPiece, sizeof(uint8_t) * pieceLen);
         }
         return redraw;
 
@@ -140,7 +140,7 @@ enum boardAction handleUserInput(enum userRequest r, int *board, struct pieceInf
     return none;
 }
 
-bool checkMove(int *piece, int piecePosX, int piecePosY, enum shape pieceShape, int *boardToCheck){
+bool checkMove(uint8_t *piece, uint8_t piecePosX, uint8_t piecePosY, enum shape pieceShape, int *boardToCheck){
 
     int sideLen = sideLenByShape(pieceShape);
 
@@ -170,17 +170,17 @@ bool checkMove(int *piece, int piecePosX, int piecePosY, enum shape pieceShape, 
     return true;
 }
 
-void rotateRight(int *dest, enum shape curShape){
+void rotateRight(uint8_t *dest, enum shape curShape){
 
     if(curShape == SquareShape){
         return;
     }
 
-    int tmpShapeCopy[MAX_PIECE_LEN];
+    uint8_t tmpShapeCopy[MAX_PIECE_LEN];
     int pos = 0;
     int sideLen = sideLenByShape(curShape);
 
-    memcpy(tmpShapeCopy, dest, pieceLenByShape(curShape) * sizeof(int));
+    memcpy(tmpShapeCopy, dest, pieceLenByShape(curShape) * sizeof(uint8_t));
 
     for(int x = 0; x < sideLen; x++){
     	for(int y = sideLen - 1; y >= 0; y--){
@@ -190,13 +190,13 @@ void rotateRight(int *dest, enum shape curShape){
     }
 }
 
-void rotateLeft(int *dest, enum shape curShape){
+void rotateLeft(uint8_t *dest, enum shape curShape){
 
-    int tmpShapeCopy[MAX_PIECE_LEN];
+    uint8_t tmpShapeCopy[MAX_PIECE_LEN];
     int pos = 0;
     int sideLen = sideLenByShape(curShape);
 
-    memcpy(tmpShapeCopy, dest, pieceLenByShape(curShape) * sizeof(int));
+    memcpy(tmpShapeCopy, dest, pieceLenByShape(curShape) * sizeof(uint8_t));
 
     for(int x = sideLen - 1; x >= 0; x--){
         for (int y = 0; y < sideLen; y++){
